+++ src/Document/ProjectDocument.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Symfony\Lsp\Document;
+
+final class ProjectDocument
+{
+    public function __construct(
+        public readonly string $text,
+        public readonly ?int $version,
+    ) {
+    }
+}